Refactor: #19 Separate PokemonDexDetailView
- PokemonDexDetailView 별도 파일로 분리
